requirements.txt
langchain==0.3.0
langchain-community==0.3.0
langchain-core==0.3.0
langchain-openai==0.2.0
langchain-chroma==0.1.4
langgraph==0.2.20
pypdf==5.0.0
chromadb==0.5.3
python-dotenv==1.0.1
chainlit==1.3.0
reportlab==4.2.2
tiktoken==0.7.0
pydantic==2.9.2
